requirements.txt
langchain
google-generativeai
google-cloud-aiplatform
python-dotenv
sentence-transformers
openai
pandas
cohere
numpy
streamlit